pyproject.toml
[project]
name = "recllm_indexer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cohere>=5.15.0",
    "google-genai>=1.10.0",
    "pgvector>=0.4.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "tqdm>=4.67.1",
]
